move github-app tests to feature directory (#36040)
